cpufreq: interactive: Skip timer when policy->min == policy->max
When the minfreq is set to the maxfreq, there is nothing to really do: the
CPU is already running at its maxfreq. In this scenario, just do nothing
and re-arm the timer.

This scenario occurs when the screen is turned on due to the cpu_input_boost
driver (it sets policy->min to policy->max for a few seconds).

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>
Signed-off-by: Ali Hasan <ahb7671@gmail.com>
Change-Id: I9526219b1e702eec01d45b949d10ca1663c17f4c